+++ model_server_tools/model_packaging/setup.py
@@ -73,7 +73,7 @@ if __name__ == '__main__':
             'onnx': ['onnx==1.1.1']
         },
         entry_points={
-            'console_scripts': ['model-export-tool=model_server_util_tools.model_packaging.export_model:export']
+            'console_scripts': ['model-export-tool=model_server_tools.model_packaging.export_model:export']
         },
         include_package_data=True,
         license='Apache License Version 2.0'
